pyproject.toml
[project]
name = "kaianolevine-api"
version = "1.21.1"
description = "FastAPI service powering api.kaianolevine.com — sets, tracks, catalog, stats, evaluations, contact, and resume."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.27,<1.0",
  "sqlalchemy[asyncio]>=2.0,<3.0",
  "asyncpg>=0.29,<1.0",
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "httpx>=0.27,<1.0",
  "cryptography>=42.0,<44.0",
  "sentry-sdk[fastapi]>=2.0,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "common-python-utils @ git+https://github.com/mini-app-polis/common-python-utils.git@main",
  "pyjwt[crypto]>=2.8,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "pytest-cov>=5.0,<6.0",
  "respx>=0.21,<1.0",
  "aiosqlite>=0.20,<1.0",
  "pre-commit>=3.6,<4.0",
  "ruff>=0.7,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kaianolevine_api"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.uv.sources]
common-python-utils = { git = "https://github.com/mini-app-polis/common-python-utils.git", rev = "main" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
